Move redirect_to_signin helper to new util module
We'll expand this module in later commits.
