+++ app/main/views/code_not_received.py
@@ -4,7 +4,7 @@ from app import user_api_client
 from app.main import main
 from app.main.forms import TextNotReceivedForm
 from app.models.user import User
-from app.utils import redirect_to_sign_in
+from app.utils.login import redirect_to_sign_in
 
 
 @main.route('/resend-email-verification')
