+++ app/utils/login.py
@@ -1,0 +1,13 @@
+from functools import wraps
+
+from flask import redirect, session, url_for
+
+
+def redirect_to_sign_in(f):
+    @wraps(f)
+    def wrapped(*args, **kwargs):
+        if 'user_details' not in session:
+            return redirect(url_for('main.sign_in'))
+        else:
+            return f(*args, **kwargs)
+    return wrapped
